+++ test/hdf5lib/TestH5A.java
@@ -22,13 +22,13 @@ public class TestH5A {
     private static final String H5_FILE = "test.h5";
     private static final int DIM_X = 4;
     private static final int DIM_Y = 6;
-    int H5fid = -1;
-    int H5dsid = -1;
-    int H5did = -1;
+    long H5fid = -1;
+    long H5dsid = -1;
+    long H5did = -1;
     long[] H5dims = { DIM_X, DIM_Y };
-    int type_id = -1;
-    int space_id = -1;
-    int lapl_id = -1;
+    long type_id = -1;
+    long space_id = -1;
+    long lapl_id = -1;
 
     private final void _deleteFile(String filename) {
         File file = new File(filename);
@@ -38,8 +38,8 @@ public class TestH5A {
         }
     }
 
-    private final int _createDataset(int fid, int dsid, String name, int dapl) {
-        int did = -1;
+    private final long _createDataset(long fid, long dsid, String name, long dapl) {
+        long did = -1;
         try {
             did = H5.H5Dcreate(fid, name, HDF5Constants.H5T_STD_I32BE, dsid,
                     HDF5Constants.H5P_DEFAULT, HDF5Constants.H5P_DEFAULT, dapl);
@@ -104,7 +104,7 @@ public class TestH5A {
     
     @Test
     public void testH5Acreate2() {
-        int attr_id = -1;
+        long attr_id = -1;
         try {
             attr_id = H5.H5Acreate(H5did, "dset", type_id, space_id, HDF5Constants.H5P_DEFAULT, HDF5Constants.H5P_DEFAULT);
             assertTrue("testH5Acreate2", attr_id >= 0);
@@ -132,8 +132,8 @@ public class TestH5A {
     @Test
     public void testH5Aopen() {
         String attr_name = "dset";
-        int attribute_id = -1;
-        int attr_id = -1;
+        long attribute_id = -1;
+        long attr_id = -1;
 
         try {
             attr_id = H5.H5Acreate(H5did, attr_name, type_id, space_id,
@@ -164,14 +164,14 @@ public class TestH5A {
 
     @Test
     public void testH5Aopen_by_idx() {
-        int loc_id = H5did;
+        long loc_id = H5did;
         String obj_name = ".";
         int idx_type = HDF5Constants.H5_INDEX_CRT_ORDER;
         int order = HDF5Constants.H5_ITER_INC;
         long n = 0;
-        int attr_id = -1;
-        int attribute_id = -1;
-        int aapl_id = HDF5Constants.H5P_DEFAULT;
+        long attr_id = -1;
+        long attribute_id = -1;
+        long aapl_id = HDF5Constants.H5P_DEFAULT;
 
         try {
             attr_id = H5.H5Acreate(H5did, "file", type_id, space_id,
@@ -229,7 +229,7 @@ public class TestH5A {
     public void testH5Acreate_by_name() {
         String obj_name = ".";
         String attr_name = "DATASET";
-        int attribute_id = -1;
+        long attribute_id = -1;
         boolean bool_val = false;
 
         try {
@@ -258,10 +258,10 @@ public class TestH5A {
     
     @Test
     public void testH5Arename() throws Throwable, HDF5LibraryException, NullPointerException {
-        int loc_id = H5fid;
+        long loc_id = H5fid;
         String old_attr_name = "old";
         String new_attr_name = "new";
-        int attr_id = -1;
+        long attr_id = -1;
         int ret_val = -1;
         boolean bool_val = false;
 
@@ -295,11 +295,11 @@ public class TestH5A {
     
     @Test
     public void testH5Arename_by_name() {
-        int loc_id = H5fid;
+        long loc_id = H5fid;
         String obj_name = ".";
         String old_attr_name = "old";
         String new_attr_name = "new";
-        int attr_id = -1;
+        long attr_id = -1;
         int ret_val = -1;
         boolean bool_val = false;
 
@@ -338,15 +338,15 @@ public class TestH5A {
     
     @Test
     public void testH5Aget_name_by_idx() {
-        int loc_id = H5fid;
+        long loc_id = H5fid;
         String obj_name = ".";
         String attr_name = "DATASET1", attr2_name = "DATASET2";
         String ret_name = null;
         int idx_type = HDF5Constants.H5_INDEX_NAME;
         int order = HDF5Constants.H5_ITER_INC;
         int n = 0;
-        int attr1_id = -1;
-        int attr2_id = -1;
+        long attr1_id = -1;
+        long attr2_id = -1;
 
         try {
             attr1_id = H5.H5Acreate_by_name(loc_id, obj_name, attr_name,
@@ -383,7 +383,7 @@ public class TestH5A {
 
     @Test
     public void testH5Aget_storage_size() {
-        int attr_id = -1;
+        long attr_id = -1;
         long attr_size = -1;
 
         try {
@@ -406,8 +406,8 @@ public class TestH5A {
     @Test
     public void testH5Aget_info() {
         H5A_info_t attr_info = null;
-        int attribute_id = -1;
-        int attr_id = -1;
+        long attribute_id = -1;
+        long attr_id = -1;
 
         try {
             attr_id = H5.H5Acreate(H5did, "dset", type_id, space_id,
@@ -438,8 +438,8 @@ public class TestH5A {
     @Test
     public void testH5Aget_info1() {
         H5A_info_t attr_info = null;
-        int attribute_id = -1;
-        int attr_id = -1;
+        long attribute_id = -1;
+        long attr_id = -1;
         int order = HDF5Constants.H5_ITER_INC;
 
         try {
@@ -474,8 +474,8 @@ public class TestH5A {
     
     @Test
     public void testH5Aget_info_by_idx() {
-        int attr_id = -1;
-        int attr2_id = -1;;
+        long attr_id = -1;
+        long attr2_id = -1;;
         H5A_info_t attr_info = null;
 
         try {
@@ -530,7 +530,7 @@ public class TestH5A {
 
     @Test
     public void testH5Aget_info_by_name() {
-        int attr_id = -1;
+        long attr_id = -1;
         H5A_info_t attr_info = null;
         String obj_name = ".";
         String attr_name = "DATASET";
@@ -555,7 +555,7 @@ public class TestH5A {
     
     @Test
     public void testH5Adelete_by_name() {
-        int attr_id = -1;
+        long attr_id = -1;
         int ret_val = -1;
         boolean bool_val = false;
         boolean exists = false;
@@ -598,8 +598,8 @@ public class TestH5A {
     @Test
     public void testH5Aexists() {
         boolean exists = false;
-        int attr_id = -1;
-        int attribute_id = -1;
+        long attr_id = -1;
+        long attribute_id = -1;
 
         try {
             exists = H5.H5Aexists(H5fid, "None");
@@ -637,10 +637,10 @@ public class TestH5A {
     @Test
     public void testH5Adelete_by_idx_order() {
         boolean exists = false;
-        int attr1_id = -1;
-        int attr2_id = -1;
-        int attr3_id = -1;
-        int attr4_id = -1;
+        long attr1_id = -1;
+        long attr2_id = -1;
+        long attr3_id = -1;
+        long attr4_id = -1;
         
         try {
             attr1_id = H5.H5Acreate_by_name(H5fid, ".", "attribute1",
@@ -679,9 +679,9 @@ public class TestH5A {
     @Test
     public void testH5Adelete_by_idx_name1() {
         boolean exists = false;
-        int attr1_id = -1;
-        int attr2_id = -1;
-        int attr3_id = -1;
+        long attr1_id = -1;
+        long attr2_id = -1;
+        long attr3_id = -1;
         
         try {
             attr1_id = H5.H5Acreate_by_name(H5fid, ".", "attribute1",
@@ -714,10 +714,10 @@ public class TestH5A {
     @Test
     public void testH5Adelete_by_idx_name2() {
         boolean exists = false;
-        int attr1_id = -1;
-        int attr2_id = -1;
-        int attr3_id = -1;
-        int attr4_id = -1;
+        long attr1_id = -1;
+        long attr2_id = -1;
+        long attr3_id = -1;
+        long attr4_id = -1;
         
         try {
             attr1_id = H5.H5Acreate_by_name(H5fid, ".", "attribute1",
@@ -769,8 +769,8 @@ public class TestH5A {
     public void testH5Aopen_by_name() {
         String obj_name = ".";
         String attr_name = "DATASET";
-        int attribute_id = -1;
-        int aid = -1;
+        long attribute_id = -1;
+        long aid = -1;
 
         try {
             attribute_id = H5.H5Acreate_by_name(H5fid, obj_name, attr_name,
@@ -804,9 +804,9 @@ public class TestH5A {
     @Test
     public void testH5Awrite_readVL() {
         String attr_name = "VLdata";
-        int attr_id = -1;
-        int atype_id = -1;
-        int aspace_id = -1;
+        long attr_id = -1;
+        long atype_id = -1;
+        long aspace_id = -1;
         String[] str_data = { "Parting", "is such", "sweet", "sorrow." };
         long[] dims = { str_data.length };
         long lsize = 1;
