+++ test/hdf5lib/TestH5Dplist.java
@@ -25,10 +25,10 @@ public class TestH5Dplist {
     private static final int NDIMS = 2;
     private static final int FILLVAL = 99;
     private static final int RANK = 2;
-    int H5fid = -1;
-    int H5dsid = -1;
-    int H5did = -1;
-    int H5dcpl_id = -1;
+    long H5fid = -1;
+    long H5dsid = -1;
+    long H5did = -1;
+    long H5dcpl_id = -1;
     long[] H5dims = { DIM_X, DIM_Y };
     long[] H5extdims = { EDIM_X, EDIM_Y };
     long[] H5chunk_dims = { CHUNK_X, CHUNK_Y };
@@ -42,7 +42,7 @@ public class TestH5Dplist {
         }
     }
 
-    private final void _createPDataset(int fid, int dsid, String name, int dcpl_val) {
+    private final void _createPDataset(long fid, long dsid, String name, long dcpl_val) {
         try {
             H5dcpl_id = H5.H5Pcreate(dcpl_val);
         }
@@ -82,7 +82,7 @@ public class TestH5Dplist {
         _createDataset(H5fid, H5dsid, "dset", H5dcpl_id, HDF5Constants.H5P_DEFAULT);
     }
 
-    private final void _createDataset(int fid, int dsid, String name, int dcpl, int dapl) {
+    private final void _createDataset(long fid, long dsid, String name, long dcpl, long dapl) {
         try {
             H5did = H5.H5Dcreate(fid, name,
                         HDF5Constants.H5T_STD_I32BE, dsid,
